# backend-api/api/service/openapi.yaml
openapi: 3.0.3
info:
  title: WAS API
  version: "1.0.0"
...
paths:
...
  /api/v1/resources:
    get:
      operationId: ListResources
      tags:
        - resources
      summary: List resources
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resource"
    post:
      operationId: CreateResource
      tags:
        - resources
      summary: Create resource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourceCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
  /api/v1/resources/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
    get:
      operationId: GetResource
      tags:
        - resources
      summary: Get resource by id
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        "404":
          description: Not found
    put:
      operationId: UpdateResource
      tags:
        - resources
      summary: Update resource by id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourceUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        "404":
          description: Not found
    delete:
      operationId: DeleteResource
      tags:
        - resources
      summary: Delete resource by id
      responses:
        "204":
          description: No content
        "404":
          description: Not found
...
  /api/v1/ml-models:
    get:
      operationId: ListMLModels
      tags:
        - ml-models
      summary: List ML models
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MLModel"
    post:
      operationId: CreateMLModel
      tags:
        - ml-models
      summary: Create ML model
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MLModelCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MLModel"
  /api/v1/ml-models/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
    get:
      operationId: GetMLModel
      tags:
        - ml-models
      summary: Get ML model by id
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MLModel"
        "404":
          description: Not found
    put:
      operationId: UpdateMLModel
      tags:
        - ml-models
      summary: Update ML model by id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MLModelUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MLModel"
        "404":
          description: Not found
    delete:
      operationId: DeleteMLModel
      tags:
        - ml-models
      summary: Delete ML model by id
      responses:
        "204":
          description: No content
        "404":
          description: Not found
...
  /api/v1/request-logs/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
    get:
      operationId: GetRequestLog
      tags:
        - request-logs
      summary: Get request log by id
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestLog"
        "404":
          description: Not found
...
  /api/v1/event-logs/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
    get:
      operationId: GetEventLog
      tags:
        - event-logs
      summary: Get event log by id
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventLog"
        "404":
          description: Not found
components:
  schemas:
    Resource:
      type: object
      required:
        - id
        - name
        - url_pattern
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        url_pattern:
          type: string
        security_profile_id:
          type: integer
          format: int64
          nullable: true
        traffic_profile_id:
          type: integer
          format: int64
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ResourceCreate:
      type: object
      required:
        - name
        - url_pattern
      properties:
        name:
          type: string
        url_pattern:
          type: string
        security_profile_id:
          type: integer
          format: int64
          nullable: true
        traffic_profile_id:
          type: integer
          format: int64
          nullable: true
    ResourceUpdate:
      type: object
      required:
        - name
        - url_pattern
      properties:
        name:
          type: string
        url_pattern:
          type: string
        security_profile_id:
          type: integer
          format: int64
          nullable: true
        traffic_profile_id:
          type: integer
          format: int64
          nullable: true
...
    MLModel:
      type: object
      required:
        - id
        - name
        - model_data
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        model_data:
          type: string
          format: byte
    MLModelCreate:
      type: object
      required:
        - name
        - model_data
      properties:
        name:
          type: string
        model_data:
          type: string
          format: byte
    MLModelUpdate:
      type: object
      required:
        - name
        - model_data
      properties:
        name:
          type: string
        model_data:
          type: string
          format: byte
    RequestLog:
      type: object
      required:
        - id
        - occurred_at
        - client_ip
        - method
        - path
        - status_code
        - action
      properties:
        id:
          type: integer
          format: int64
        resource_id:
          type: integer
          format: int64
          nullable: true
        occurred_at:
          type: string
          format: date-time
        client_ip:
          type: string
        method:
          type: string
        path:
          type: string
        status_code:
          type: integer
          format: int32
        action:
          type: string
        rule_id:
          type: integer
          format: int64
        profile_id:
          type: integer
          format: int64
        user_agent:
          type: string
        country:
          type: string
        latency_ms:
          type: integer
          format: int32
        request_id:
          type: string
        metadata:
          type: object
          additionalProperties: true
          nullable: true
        host:
          type: string
        scheme:
          type: string
        protocol:
          type: string
        authority:
          type: string
        query:
          type: string
        source_port:
          type: integer
          format: int32
        destination_ip:
          type: string
        destination_port:
          type: integer
          format: int32
        source_principal:
          type: string
        source_service:
          type: string
        source_labels:
          type: object
          additionalProperties: true
          nullable: true
        destination_service:
          type: string
        destination_labels:
          type: object
          additionalProperties: true
          nullable: true
        request_http_id:
          type: string
        fragment:
          type: string
        request_headers:
          type: object
          additionalProperties: true
          nullable: true
        request_body_size:
          type: integer
          format: int32
        request_body:
          type: string
        context_extensions:
          type: object
          additionalProperties: true
          nullable: true
        metadata_context:
          type: object
          additionalProperties: true
          nullable: true
        route_metadata_context:
          type: object
          additionalProperties: true
          nullable: true
    EventLog:
      type: object
      required:
        - id
        - resource_id
        - occurred_at
        - event_type
        - severity
        - message
      properties:
        id:
          type: integer
          format: int64
        resource_id:
          type: integer
          format: int64
        occurred_at:
          type: string
          format: date-time
        event_type:
          type: string
        severity:
          type: string
        message:
          type: string
        rule_id:
          type: integer
          format: int64
        profile_id:
          type: integer
          format: int64
        metadata:
          type: object
          additionalProperties: true
          nullable: true